A New Beginning
This brainfuck interpreter doesn't have any optimizations and is a work
in progress. At the moment, an intermediate representation is required
to simplify programs in order to execute longer ones such as the
infamous `mandelbrot.b` in reasonable time.

A means of inputting a single character needs to be selected, as Rust
doesn't currently have a stable implementation. The preprocessing stage
currently generates a jump table so that the bracket commands do not
require deep recursion to be implemented. In the future, it will
condense operations like CLEAR (`[-]`) and MULTIPLY (`+n[>+n<-]`) and
others.

The name of this project is Faust, an allusion to the diabollically low
level operations provided by brainfuck, as well as a close homonym of
'fast', which this project intends to be. The benchmark target for
completion is the execution of mandelbrot.b within 10 seconds on the
interpreter.