Timing Code Moved
The timing system has been moved to its own module and reimplemented as
macros that wrap a block of code.

- `benchmark!{ ... }` times the code it contains and outputs simple timing
messages.
- `named_benchmark!{ expr, ... }` does the same, but allows the
benchmark to be disambiguated from nested benchmarking you may perform.
Because it takes an expr as an argument, you may use any type of value
to name the benchmark, so long as it can be printed. This allows
benchmarks to be given numerical names, as well as to show the state of
some variable when the benchmark begins.