Timing Refactor
To avoid syntax errors, the timing code now captures block fragments. It
now returns the elapsed time rather than printing. This makes it much
easier to choose how to disambiguate or process different benchmarks.
`named_benchmark!` is thus no longer necessary.

There is also a benchmark that gets the average of multiple runs, in
order to reduce sampling error. In the future, this may return a
structure containing statistical deviation, the quartiles (including the
average), the range of values, and perhaps the vector of samples for
further processing.