Major Refactor and Simple Compilation
lexer and optimizer interfaces have been removed in favor of an
interface called Frontend, which exposes a function to compile a string
to Faust intermediate representation. This allows frontends to be very
simple, or to be extraordinarily complex systems.

Many enums were moved and refactored in order to allow a more clear
layout with less interdependency. `hanoi.b` has been added to the
examples for its remarkable value as a benchmarking code. It takes a
while to execute, but outputs very pleasant graphics that change over
time. Perhaps it may be possible to make a small maze game in brainfuck
using ANSI sequences.

`mod interpreter` will be the working directory for the faust
interpreter until it is stable enough to replace the current interpreter
loop that naively processes instructions. The condense operation is the
first major optimization milestone and `mandelbrot.b` is compiled for
validation purposes. It works so far!

Next up, non-recursively simplifying loops.