Optimization Refactor
Further improvements to the frontend interface, which now allows direct
compilation and optimized compilation. Given the `basic` implementation
of a `frontend`, an optimizing version can now be accessed with
`optimize`, which takes the basic output and sends it through two tiers
of optimizations.

1. Cull a few non-printing characters
from the end.
2. Condense multiple calls to a simple operation into one command.

There is also support for single-stream mode, using the `!` character to
toggle lexing. A simple single-stream program may be implemented that
embeds a string, then reads it out:
```[Hello World - single stream]
!Hello World! ;Make a buffer
+[>,.]        ;Consume and print characters```

The next optimizations to make are loop blocks. The interpreter may now
be repurposed to consume `Vec<FaustCmd>`, and should in theory run faster
than the raw character interpreter.