Licensed as Apache 2.0
You may freely distribute, use, and modify this project. A copy of this license must accompany any redistribution or modification of this project (particularly if source is made available). You may not receive a patent and then proceed to sue for patent infringement solely on the grounds that this project has been used.