Switch to Faust Compiled Code
The interpreter now runs the compiled and optimized version of brainfuck
code. I was absolutely shocked that the performance jumped so massively
simply because of condensing repeated operations! Both hanoi.b and
mandelbrot.b execute in under a minute and profiling shows that there
was a performance increase of 1000x!

I cleaned up the old dead fragments. Next up is loop optimizations and
perhaps a nicer command line interface to separate input and output a
bit more from the implementation. This will allow the input buffer
command to do some work.

This interpreter just went from being a toy to being usable.