Optimization: Clear Cell
The clear cell command is put together to see how much improvement it
makes on execution speed. Performance is approximately the same, but an
average of 5-10ms slower now. This indicates that the switching
statement's growth in size created a few more cache misses than the
performance improvement. I'll probably see better performance when
multiplication and division commands exist.

Benchmarking was done from the test function, allowing capture of the
interpreting only. Not just compilation. This helps to compare execution
speed between versions rather than overall speed.

A few variables were renamed, particularly instances of 'instructions'
in favour of the shorter and easier to abbreviate 'commands'.

As this project has grown, it has become clear that two things will be
necessary to enable to loop-based optimizations. First, unwrapping the
enum is becoming very difficult so I'll need to switch to a struct to
carry arguments. Secondly, the concept of a 'jumptable' needs to be
generalized into a concept of loop fragments in order to reduce
duplication when it becomes necessary to analyze looping constructs like
multiplication.